+++ packages/in_app_purchase/in_app_purchase/example/lib/main.dart
@@ -284,8 +284,8 @@ class _MyAppState extends State<_MyApp> {
                           changeSubscriptionParam: (oldSubscription != null)
                               ? ChangeSubscriptionParam(
                                   oldPurchaseDetails: oldSubscription,
-                                  prorationMode:
-                                      ProrationMode.immediateWithTimeProration,
+                                  replacementMode:
+                                      ReplacementMode.withTimeProration,
                                 )
                               : null);
                     } else {