+++ packages/in_app_purchase/in_app_purchase_android/android/src/test/java/io/flutter/plugins/inapppurchase/BillingClientFactoryImplTest.java
@@ -55,7 +55,7 @@ public class BillingClientFactoryImplTest {
     // No logic to verify just ensure creation works.
     BillingClient client =
         factory.createBillingClient(
-            context, mockCallbackApi, PlatformBillingChoiceMode.PLAY_BILLING_ONLY);
+            context, mockCallbackApi, PlatformBillingChoiceMode.PLAY_BILLING_ONLY, null);
     assertNotNull(client);
   }
 
@@ -64,7 +64,7 @@ public class BillingClientFactoryImplTest {
     // No logic to verify just ensure creation works.
     BillingClient client =
         factory.createBillingClient(
-            context, mockCallbackApi, PlatformBillingChoiceMode.ALTERNATIVE_BILLING_ONLY);
+            context, mockCallbackApi, PlatformBillingChoiceMode.ALTERNATIVE_BILLING_ONLY, null);
     assertNotNull(client);
   }
 
@@ -78,7 +78,7 @@ public class BillingClientFactoryImplTest {
 
     final BillingClient billingClient =
         factory.createBillingClient(
-            context, mockCallbackApi, PlatformBillingChoiceMode.USER_CHOICE_BILLING);
+            context, mockCallbackApi, PlatformBillingChoiceMode.USER_CHOICE_BILLING, null);
 
     UserChoiceDetails details = mock(UserChoiceDetails.class);
     final String externalTransactionToken = "someLongTokenId1234";
@@ -96,6 +96,17 @@ public class BillingClientFactoryImplTest {
     assertEquals(
         callbackCaptor.getValue().getOriginalExternalTransactionId(), originalTransactionId);
     assertTrue(callbackCaptor.getValue().getProducts().isEmpty());
+  }
+
+  @Test
+  public void pendingPurchasesForPrepaidPlans() {
+    // No logic to verify just ensure creation works.
+    Messages.PendingPurchasesParams params =
+        new Messages.PendingPurchasesParams.Builder().setEnablePrepaidPlans(true).build();
+    BillingClient client =
+        factory.createBillingClient(
+            context, mockCallbackApi, PlatformBillingChoiceMode.PLAY_BILLING_ONLY, params);
+    assertNotNull(client);
   }
 
   @After