+++ packages/in_app_purchase/in_app_purchase_android/android/src/test/java/io/flutter/plugins/inapppurchase/MethodCallHandlerTest.java
@@ -5,7 +5,6 @@
 package io.flutter.plugins.inapppurchase;
 
 import static io.flutter.plugins.inapppurchase.MethodCallHandlerImpl.ACTIVITY_UNAVAILABLE;
-import static io.flutter.plugins.inapppurchase.MethodCallHandlerImpl.PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY;
 import static io.flutter.plugins.inapppurchase.MethodCallHandlerImpl.REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY;
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
@@ -104,22 +103,33 @@ public class MethodCallHandlerTest {
   @Mock Context context;
   @Mock ActivityPluginBinding mockActivityPluginBinding;
 
+  private final Messages.PendingPurchasesParams defaultPendingPurchasesParams =
+      new Messages.PendingPurchasesParams.Builder().setEnablePrepaidPlans(false).build();
+
   private final Long DEFAULT_HANDLE = 1L;
 
   @Before
   public void setUp() {
     openMocks = MockitoAnnotations.openMocks(this);
+
     // Use the same client no matter if alternative billing is enabled or not.
     when(factory.createBillingClient(
-            context, mockCallbackApi, PlatformBillingChoiceMode.PLAY_BILLING_ONLY))
+            context,
+            mockCallbackApi,
+            PlatformBillingChoiceMode.PLAY_BILLING_ONLY,
+            defaultPendingPurchasesParams))
         .thenReturn(mockBillingClient);
     when(factory.createBillingClient(
-            context, mockCallbackApi, PlatformBillingChoiceMode.ALTERNATIVE_BILLING_ONLY))
+            context,
+            mockCallbackApi,
+            PlatformBillingChoiceMode.ALTERNATIVE_BILLING_ONLY,
+            defaultPendingPurchasesParams))
         .thenReturn(mockBillingClient);
     when(factory.createBillingClient(
             any(Context.class),
             any(InAppPurchaseCallbackApi.class),
-            eq(PlatformBillingChoiceMode.USER_CHOICE_BILLING)))
+            eq(PlatformBillingChoiceMode.USER_CHOICE_BILLING),
+            any(Messages.PendingPurchasesParams.class)))
         .thenReturn(mockBillingClient);
     methodChannelHandler = new MethodCallHandlerImpl(activity, context, mockCallbackApi, factory);
     when(mockActivityPluginBinding.getActivity()).thenReturn(activity);
@@ -159,10 +169,15 @@ public class MethodCallHandlerTest {
   @Test
   public void startConnection() {
     ArgumentCaptor<BillingClientStateListener> captor =
-        mockStartConnection(PlatformBillingChoiceMode.PLAY_BILLING_ONLY);
+        mockStartConnection(
+            PlatformBillingChoiceMode.PLAY_BILLING_ONLY, defaultPendingPurchasesParams);
     verify(platformBillingResult, never()).success(any());
     verify(factory, times(1))
-        .createBillingClient(context, mockCallbackApi, PlatformBillingChoiceMode.PLAY_BILLING_ONLY);
+        .createBillingClient(
+            context,
+            mockCallbackApi,
+            PlatformBillingChoiceMode.PLAY_BILLING_ONLY,
+            defaultPendingPurchasesParams);
 
     BillingResult billingResult = buildBillingResult();
     captor.getValue().onBillingSetupFinished(billingResult);
@@ -177,11 +192,15 @@ public class MethodCallHandlerTest {
   @Test
   public void startConnectionAlternativeBillingOnly() {
     ArgumentCaptor<BillingClientStateListener> captor =
-        mockStartConnection(PlatformBillingChoiceMode.ALTERNATIVE_BILLING_ONLY);
+        mockStartConnection(
+            PlatformBillingChoiceMode.ALTERNATIVE_BILLING_ONLY, defaultPendingPurchasesParams);
     verify(platformBillingResult, never()).success(any());
     verify(factory, times(1))
         .createBillingClient(
-            context, mockCallbackApi, PlatformBillingChoiceMode.ALTERNATIVE_BILLING_ONLY);
+            context,
+            mockCallbackApi,
+            PlatformBillingChoiceMode.ALTERNATIVE_BILLING_ONLY,
+            defaultPendingPurchasesParams);
 
     BillingResult billingResult = buildBillingResult();
     captor.getValue().onBillingSetupFinished(billingResult);
@@ -194,16 +213,40 @@ public class MethodCallHandlerTest {
   }
 
   @Test
+  public void startConnectionPendingPurchasesPrepaidPlans() {
+    Messages.PendingPurchasesParams pendingPurchasesParams =
+        new Messages.PendingPurchasesParams.Builder().setEnablePrepaidPlans(true).build();
+    ArgumentCaptor<BillingClientStateListener> captor =
+        mockStartConnection(PlatformBillingChoiceMode.USER_CHOICE_BILLING, pendingPurchasesParams);
+    verify(platformBillingResult, never()).success(any());
+    verify(factory, times(1))
+        .createBillingClient(
+            context,
+            mockCallbackApi,
+            PlatformBillingChoiceMode.USER_CHOICE_BILLING,
+            pendingPurchasesParams);
+
+    BillingResult billingResult = buildBillingResult();
+    captor.getValue().onBillingSetupFinished(billingResult);
+
+    ArgumentCaptor<PlatformBillingResult> resultCaptor =
+        ArgumentCaptor.forClass(PlatformBillingResult.class);
+    verify(platformBillingResult, times(1)).success(resultCaptor.capture());
+  }
+
+  @Test
   public void startConnectionUserChoiceBilling() {
     ArgumentCaptor<BillingClientStateListener> captor =
-        mockStartConnection(PlatformBillingChoiceMode.USER_CHOICE_BILLING);
+        mockStartConnection(
+            PlatformBillingChoiceMode.USER_CHOICE_BILLING, defaultPendingPurchasesParams);
     verify(platformBillingResult, never()).success(any());
 
     verify(factory, times(1))
         .createBillingClient(
             any(Context.class),
             any(InAppPurchaseCallbackApi.class),
-            eq(PlatformBillingChoiceMode.USER_CHOICE_BILLING));
+            eq(PlatformBillingChoiceMode.USER_CHOICE_BILLING),
+            any(Messages.PendingPurchasesParams.class));
 
     BillingResult billingResult =
         BillingResult.newBuilder()
@@ -221,10 +264,15 @@ public class MethodCallHandlerTest {
   public void userChoiceBillingOnSecondConnection() {
     // First connection.
     ArgumentCaptor<BillingClientStateListener> captor1 =
-        mockStartConnection(PlatformBillingChoiceMode.PLAY_BILLING_ONLY);
+        mockStartConnection(
+            PlatformBillingChoiceMode.PLAY_BILLING_ONLY, defaultPendingPurchasesParams);
     verify(platformBillingResult, never()).success(any());
     verify(factory, times(1))
-        .createBillingClient(context, mockCallbackApi, PlatformBillingChoiceMode.PLAY_BILLING_ONLY);
+        .createBillingClient(
+            context,
+            mockCallbackApi,
+            PlatformBillingChoiceMode.PLAY_BILLING_ONLY,
+            defaultPendingPurchasesParams);
 
     BillingResult billingResult1 =
         BillingResult.newBuilder()
@@ -248,13 +296,15 @@ public class MethodCallHandlerTest {
 
     // Second connection.
     ArgumentCaptor<BillingClientStateListener> captor2 =
-        mockStartConnection(PlatformBillingChoiceMode.USER_CHOICE_BILLING);
+        mockStartConnection(
+            PlatformBillingChoiceMode.USER_CHOICE_BILLING, defaultPendingPurchasesParams);
     verify(platformBillingResult, never()).success(any());
     verify(factory, times(1))
         .createBillingClient(
             any(Context.class),
             any(InAppPurchaseCallbackApi.class),
-            eq(PlatformBillingChoiceMode.USER_CHOICE_BILLING));
+            eq(PlatformBillingChoiceMode.USER_CHOICE_BILLING),
+            eq(defaultPendingPurchasesParams));
 
     BillingResult billingResult2 =
         BillingResult.newBuilder()
@@ -273,7 +323,10 @@ public class MethodCallHandlerTest {
     doNothing().when(mockBillingClient).startConnection(captor.capture());
 
     methodChannelHandler.startConnection(
-        DEFAULT_HANDLE, PlatformBillingChoiceMode.PLAY_BILLING_ONLY, platformBillingResult);
+        DEFAULT_HANDLE,
+        PlatformBillingChoiceMode.PLAY_BILLING_ONLY,
+        defaultPendingPurchasesParams,
+        platformBillingResult);
     verify(platformBillingResult, never()).success(any());
     BillingResult billingResult1 = buildBillingResult();
     BillingResult billingResult2 =
@@ -480,7 +533,10 @@ public class MethodCallHandlerTest {
     @SuppressWarnings("unchecked")
     final Messages.Result<PlatformBillingResult> mockResult = mock(Messages.Result.class);
     methodChannelHandler.startConnection(
-        disconnectCallbackHandle, PlatformBillingChoiceMode.PLAY_BILLING_ONLY, mockResult);
+        disconnectCallbackHandle,
+        PlatformBillingChoiceMode.PLAY_BILLING_ONLY,
+        defaultPendingPurchasesParams,
+        mockResult);
     final BillingClientStateListener stateListener = captor.getValue();
 
     // Disconnect the connected client
@@ -555,8 +611,6 @@ public class MethodCallHandlerTest {
     queryForProducts(singletonList(productId));
     PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
     paramsBuilder.setProduct(productId);
-    paramsBuilder.setProrationMode(
-        (long) PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
     paramsBuilder.setReplacementMode(
         (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
 
@@ -582,8 +636,6 @@ public class MethodCallHandlerTest {
     PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
-    paramsBuilder.setProrationMode(
-        (long) PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
     paramsBuilder.setReplacementMode(
         (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
 
@@ -613,8 +665,6 @@ public class MethodCallHandlerTest {
     PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
-    paramsBuilder.setProrationMode(
-        (long) PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
     paramsBuilder.setReplacementMode(
         (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
 
@@ -638,8 +688,6 @@ public class MethodCallHandlerTest {
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
     paramsBuilder.setOldProduct(oldProductId);
-    paramsBuilder.setProrationMode(
-        (long) PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
     paramsBuilder.setReplacementMode(
         (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
 
@@ -667,8 +715,6 @@ public class MethodCallHandlerTest {
     PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
-    paramsBuilder.setProrationMode(
-        (long) PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
     paramsBuilder.setReplacementMode(
         (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
 
@@ -687,27 +733,22 @@ public class MethodCallHandlerTest {
     assertResultsMatch(platformResult, billingResult);
   }
 
-  // TODO(gmackall): Replace uses of deprecated ProrationMode enum values with new
-  // ReplacementMode enum values.
-  // https://github.com/flutter/flutter/issues/128957.
   @Test
-  @SuppressWarnings(value = "deprecation")
   public void launchBillingFlow_ok_Proration() {
     // Fetch the product details first and query the method call
     String productId = "foo";
     String oldProductId = "oldFoo";
     String purchaseToken = "purchaseTokenFoo";
     String accountId = "account";
-    int prorationMode = BillingFlowParams.ProrationMode.IMMEDIATE_AND_CHARGE_PRORATED_PRICE;
+    int replacementMode =
+        BillingFlowParams.SubscriptionUpdateParams.ReplacementMode.CHARGE_PRORATED_PRICE;
     queryForProducts(unmodifiableList(asList(productId, oldProductId)));
     PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
     paramsBuilder.setOldProduct(oldProductId);
     paramsBuilder.setPurchaseToken(purchaseToken);
-    paramsBuilder.setProrationMode((long) prorationMode);
-    paramsBuilder.setReplacementMode(
-        (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
+    paramsBuilder.setReplacementMode((long) replacementMode);
 
     // Launch the billing flow
     BillingResult billingResult = buildBillingResult();
@@ -724,25 +765,20 @@ public class MethodCallHandlerTest {
     assertResultsMatch(platformResult, billingResult);
   }
 
-  // TODO(gmackall): Replace uses of deprecated ProrationMode enum values with new
-  // ReplacementMode enum values.
-  // https://github.com/flutter/flutter/issues/128957.
   @Test
-  @SuppressWarnings(value = "deprecation")
   public void launchBillingFlow_ok_Proration_with_null_OldProduct() {
     // Fetch the product details first and query the method call
     String productId = "foo";
     String accountId = "account";
     String queryOldProductId = "oldFoo";
-    int prorationMode = BillingFlowParams.ProrationMode.IMMEDIATE_AND_CHARGE_PRORATED_PRICE;
+    int replacementMode =
+        BillingFlowParams.SubscriptionUpdateParams.ReplacementMode.CHARGE_PRORATED_PRICE;
     queryForProducts(unmodifiableList(asList(productId, queryOldProductId)));
     PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
     paramsBuilder.setOldProduct(null);
-    paramsBuilder.setProrationMode((long) prorationMode);
-    paramsBuilder.setReplacementMode(
-        (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
+    paramsBuilder.setReplacementMode((long) replacementMode);
 
     // Launch the billing flow
     BillingResult billingResult = buildBillingResult();
@@ -773,8 +809,6 @@ public class MethodCallHandlerTest {
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
     paramsBuilder.setOldProduct(null);
-    paramsBuilder.setProrationMode(
-        (long) PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
     paramsBuilder.setReplacementMode((long) replacementMode);
 
     // Launch the billing flow
@@ -793,60 +827,21 @@ public class MethodCallHandlerTest {
   }
 
   @Test
-  @SuppressWarnings(value = "deprecation")
-  public void launchBillingFlow_ok_Proration_and_Replacement_conflict() {
-    // Fetch the product details first and query the method call
-    String productId = "foo";
-    String accountId = "account";
-    String queryOldProductId = "oldFoo";
-    int prorationMode = BillingFlowParams.ProrationMode.IMMEDIATE_AND_CHARGE_PRORATED_PRICE;
-    int replacementMode =
-        BillingFlowParams.SubscriptionUpdateParams.ReplacementMode.CHARGE_PRORATED_PRICE;
-    queryForProducts(unmodifiableList(asList(productId, queryOldProductId)));
-    PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
-    paramsBuilder.setProduct(productId);
-    paramsBuilder.setAccountId(accountId);
-    paramsBuilder.setOldProduct(queryOldProductId);
-    paramsBuilder.setProrationMode((long) prorationMode);
-    paramsBuilder.setReplacementMode((long) replacementMode);
-
-    // Launch the billing flow
-    BillingResult billingResult = buildBillingResult();
-    when(mockBillingClient.launchBillingFlow(any(), any())).thenReturn(billingResult);
-
-    // Assert that the synchronous call throws an exception.
-    FlutterError exception =
-        assertThrows(
-            FlutterError.class,
-            () -> methodChannelHandler.launchBillingFlow(paramsBuilder.build()));
-    assertEquals("IN_APP_PURCHASE_CONFLICT_PRORATION_MODE_REPLACEMENT_MODE", exception.code);
-    assertTrue(
-        Objects.requireNonNull(exception.getMessage())
-            .contains(
-                "launchBillingFlow failed because you provided both prorationMode and replacementMode. You can only provide one of them."));
-  }
-
-  // TODO(gmackall): Replace uses of deprecated ProrationMode enum values with new
-  // ReplacementMode enum values.
-  // https://github.com/flutter/flutter/issues/128957.
-  @Test
-  @SuppressWarnings(value = "deprecation")
   public void launchBillingFlow_ok_Full() {
     // Fetch the product details first and query the method call
     String productId = "foo";
     String oldProductId = "oldFoo";
     String purchaseToken = "purchaseTokenFoo";
     String accountId = "account";
-    int prorationMode = BillingFlowParams.ProrationMode.IMMEDIATE_AND_CHARGE_FULL_PRICE;
+    int replacementMode =
+        BillingFlowParams.SubscriptionUpdateParams.ReplacementMode.CHARGE_FULL_PRICE;
     queryForProducts(unmodifiableList(asList(productId, oldProductId)));
     PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
     paramsBuilder.setOldProduct(oldProductId);
     paramsBuilder.setPurchaseToken(purchaseToken);
-    paramsBuilder.setProrationMode((long) prorationMode);
-    paramsBuilder.setReplacementMode(
-        (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
+    paramsBuilder.setReplacementMode((long) replacementMode);
 
     // Launch the billing flow
     BillingResult billingResult = buildBillingResult();
@@ -872,8 +867,6 @@ public class MethodCallHandlerTest {
     PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
-    paramsBuilder.setProrationMode(
-        (long) PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
     paramsBuilder.setReplacementMode(
         (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
 
@@ -895,8 +888,6 @@ public class MethodCallHandlerTest {
     PlatformBillingFlowParams.Builder paramsBuilder = new PlatformBillingFlowParams.Builder();
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
-    paramsBuilder.setProrationMode(
-        (long) PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
     paramsBuilder.setReplacementMode(
         (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
 
@@ -921,8 +912,6 @@ public class MethodCallHandlerTest {
     paramsBuilder.setProduct(productId);
     paramsBuilder.setAccountId(accountId);
     paramsBuilder.setOldProduct(oldProductId);
-    paramsBuilder.setProrationMode(
-        (long) PRORATION_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
     paramsBuilder.setReplacementMode(
         (long) REPLACEMENT_MODE_UNKNOWN_SUBSCRIPTION_UPGRADE_DOWNGRADE_POLICY);
 
@@ -988,6 +977,7 @@ public class MethodCallHandlerTest {
   }
 
   @Test
+  @SuppressWarnings(value = "deprecation")
   public void queryPurchaseHistoryAsync() {
     // Set up an established billing client and all our mocked responses
     establishConnectedBillingClient();
@@ -1015,6 +1005,7 @@ public class MethodCallHandlerTest {
   }
 
   @Test
+  @SuppressWarnings(value = "deprecation")
   public void queryPurchaseHistoryAsync_clientDisconnected() {
     methodChannelHandler.endConnection();
 
@@ -1137,7 +1128,8 @@ public class MethodCallHandlerTest {
    * <p>Defaults to play billing only which is the default.
    */
   private ArgumentCaptor<BillingClientStateListener> mockStartConnection() {
-    return mockStartConnection(PlatformBillingChoiceMode.PLAY_BILLING_ONLY);
+    return mockStartConnection(
+        PlatformBillingChoiceMode.PLAY_BILLING_ONLY, defaultPendingPurchasesParams);
   }
 
   /**
@@ -1145,12 +1137,14 @@ public class MethodCallHandlerTest {
    * Messages.Result)} with startup params.
    */
   private ArgumentCaptor<BillingClientStateListener> mockStartConnection(
-      PlatformBillingChoiceMode billingChoiceMode) {
+      PlatformBillingChoiceMode billingChoiceMode,
+      Messages.PendingPurchasesParams pendingPurchasesParams) {
     ArgumentCaptor<BillingClientStateListener> captor =
         ArgumentCaptor.forClass(BillingClientStateListener.class);
     doNothing().when(mockBillingClient).startConnection(captor.capture());
 
-    methodChannelHandler.startConnection(DEFAULT_HANDLE, billingChoiceMode, platformBillingResult);
+    methodChannelHandler.startConnection(
+        DEFAULT_HANDLE, billingChoiceMode, pendingPurchasesParams, platformBillingResult);
     return captor;
   }
 
@@ -1158,7 +1152,10 @@ public class MethodCallHandlerTest {
     @SuppressWarnings("unchecked")
     final Messages.Result<PlatformBillingResult> mockResult = mock(Messages.Result.class);
     methodChannelHandler.startConnection(
-        DEFAULT_HANDLE, PlatformBillingChoiceMode.PLAY_BILLING_ONLY, mockResult);
+        DEFAULT_HANDLE,
+        PlatformBillingChoiceMode.PLAY_BILLING_ONLY,
+        defaultPendingPurchasesParams,
+        mockResult);
   }
 
   private void queryForProducts(List<String> productIdList) {